+++ CHANGELOG.md
@@ -145,6 +145,7 @@
 - [10417](https://github.com/vegaprotocol/vega/issues/10417) - Party margin modes `API` always errors.
 - [10431](https://github.com/vegaprotocol/vega/issues/10431) - Fix source staleness validation. 
 - [10436](https://github.com/vegaprotocol/vega/issues/10436) - Fix source staleness validation when oracles are not defined.
+- [10434](https://github.com/vegaprotocol/vega/issues/10434) - Unsubscribe oracles when market is closed.
 
 ## 0.73.0